reader is not filesystem
There were only 2 functions in not.reader and only 1 was used.
`deserialize_task_spec_file` certainly _sounds_ like it'd belong to
`filesysetm`...
